requirements.txt
PySide6==6.9.1
PyYAML==6.0.2
keyboard==0.13.5
moviepy==2.2.1
imageio-ffmpeg==0.6.0
pillow==11.3.0
numpy==2.2.6
librosa==0.11.0
opencv-python==4.12.0.88
pydub==0.25.1
pytesseract==0.3.13
scipy==1.16.1
numba==0.61.2
llvmlite==0.44.0
soundfile==0.13.1
soxr==0.5.0.post1
audioread==3.0.1
tqdm==4.67.1
